feat: oneOf pointer w/o data should return schema
this is the breaking change proposal.
if a non-breaking change is preferred, I can propose a PR where we add a
`getRawSchema()` that passes a config option to `step()` to behave this
way

this is a fantastic library for the codemirror 6 mode we are building,
however we've had to
[patch](https://github.com/acao/cm6-language-json-schema/pull/2/files#diff-84b470e1f66294b6222e35877b2d17f64a790ffb6d5e862584030cc2341f31ea)
for this, which we can't ship with obviously.

what do you think?